created initial ticket box outline
